takeout control3_step
tests are failing good
