+++ contrlWithGeodesics/__init__.py
@@ -1,5 +1,12 @@
 print("Control activated and ready to use")
-__all__ = ["geodesic", "pauli_mat_vec", "fidelity", "muk", "controlSetup3"]
+__all__ = [
+    "geodesic",
+    "pauli_mat_vec",
+    "fidelity",
+    "muk",
+    "controlSetup3",
+    "control3_step",
+]
 
 from .geodesic import geodesic
 from .fidelity import fidelity
@@ -7,3 +14,4 @@ from .muk import muk
 from .pauli_mat_vec import bloch_vector
 from .controlSetup3 import control1setup3
 from .getTimeFidelity import get_time_fidelity
+from .control3_step import control3_step
