+++ contrlWithGeodesics/controlSetup3.py
@@ -5,6 +5,8 @@ import numpy as np
 from scipy.integrate import solve_ivp
 from sympy.solvers import solve
 
+from .control3_step import control3_step
+
 
 def control1setup3(
     ri, sf, Nmax=60, w0=5, gamma_0=0.01, gamma_c=10, deltat=0.003, initime=0.0
@@ -29,8 +31,11 @@ def control1setup3(
     vector_lambda = list([])
     # iterate to find the lambda value
     helperk = 0
-    # auxtime =
-    oldri = ri  # inicializar ri calculculando explicitamente con el paso siguiente
+    oldri = ri  ## save initial state
+    xri, yri, zri, solnri = control3_step(
+        ri, sf, lambda_x, w0, gamma_0, gamma_c, deltat, D_matrix, vector_lambda
+    )  # initialize the new evolve state
+    ri = np.array([xri, yri, zri])
 
     # while (fidelity(oldri, sf) <= fidelity(ri, sf)) and (helperk < Nmax):
     print("before while")
@@ -40,71 +45,15 @@ def control1setup3(
         """helperk==0: eliminarlo y ri = calcular explicitamente con el paso siguiente"""
         print("k", helperk)
         print(fidelity(ri, sf) - fidelity(oldri, sf))
-        # calculate vk, ukx, uky, and ukz
-        vk = (
-            lambda_x / 4.0 * (2.0 * muk(ri, sf)[2] - np.dot(ri, D_matrix @ muk(ri, sf)))
-        )
-        ukx = lambda_x * np.cross(ri, muk(ri, sf))[0]
-        uky = lambda_x * np.cross(ri, muk(ri, sf))[1]
-        ukz = lambda_x * np.cross(ri, muk(ri, sf))[2]
-        # print("estado initial",ri)
-        # print("estado final",sf)
-        # print("muk",muk(ri, sf))
-        # print(vk, ukx, uky, ukz)
 
-        # solve for lambda
-        solver1 = solve(
-            vk**2 + ukx**2 + uky**2 + ukz**2 - gamma_c**2, lambda_x
+        x, y, z, soln = control3_step(
+            ri, sf, lambda_x, w0, gamma_0, gamma_c, deltat, D_matrix, vector_lambda
         )
 
-        lamdasol = solver1[1]
-
-        vector_lambda.append(lamdasol)
-        auxilvk = (
-            lamdasol / 4 * (2.0 * muk(ri, sf)[2] - np.dot(ri, D_matrix @ muk(ri, sf)))
-        )
-        vk = auxilvk if auxilvk > 0 else 0
-        ukx = lamdasol * np.cross(ri, muk(ri, sf))[0]
-        uky = lamdasol * np.cross(ri, muk(ri, sf))[1]
-        ukz = lamdasol * np.cross(ri, muk(ri, sf))[2]
-
-        # Matrix for setup 3
-        Bmatrix = np.array(
-            [
-                [-2.0 * gamma_0 - vk / 2.0, -2.0 * (w0 + ukz), 2.0 * uky],
-                [2.0 * (w0 + ukz), -2.0 * gamma_0 - vk / 2.0, -2.0 * ukx],
-                [-2.0 * uky, 2.0 * ukx, -vk],
-            ]
-        )
-        qk = [0.0, 0.0, vk]  # qk setup3
-
-        def odes(t, X):
-            ## this function defines the differential equation
-            rkx, rky, rkz = X
-            drkx = (Bmatrix @ [rkx, rky, rkz] + qk)[0]
-            drky = (Bmatrix @ [rkx, rky, rkz] + qk)[1]
-            drkz = (Bmatrix @ [rkx, rky, rkz] + qk)[2]
-            return drkx, drky, drkz
-
-        ## initial conditions of the ODE
-        rkx0 = ri[0]
-        rky0 = ri[1]
-        rkz0 = ri[2]
-        # here we solve the ODE
-        soln = solve_ivp(odes, (0.0, deltat), (rkx0, rky0, rkz0), dense_output=True)
-
-        # replace to get x,y,z to plot on the Bloch sphere
-        x, y, z = soln.sol(deltat)
         oldri = ri  ## save the old ri
-        ri = np.array([x, y, z])  ## update the ri
-        # print("Just checking-the new vector is:",ri)
-        # print("Just checking-the vector to get:",sf)
-        # print("Just checking-the old fidelity is:",fidelity(oldri,sf))
-        # print("Just checking-the fidelity is:",fidelity(ri,sf))
-        # print("Just checking-helperk=",helperk)
+        ri = np.array([x, y, z])
         c.append(ri)
         initime = initime + deltat
-        # print(initime)
         tiempototal.append(initime)
         helperk += 1
     ### eliminate the final states
